Pull out some validation logic into a separate class.
